requirements.txt
# streamlit==1.38.0
# google-api-python-client==2.149.0
# groq==0.11.0
# python-dotenv==1.0.1
# pandas==2.2.3


# streamlit
# google-api-python-client
# groq
# python-dotenv
# pandas
# httpx


# streamlit==1.38.0
# google-api-python-client==2.149.0
# groq==0.11.0
# python-dotenv==1.0.1
# pandas==2.2.3
# httpx==0.27.0
# requests==2.32.3


# streamlit
# google-api-python-client
# groq
# python-dotenv
# pandas
# httpx
# requests

# streamlit==1.38.0
# google-api-python-client==2.149.0
# groq==0.11.0
# python-dotenv==1.0.1
# pandas==2.2.3
# httpx==0.27.0

streamlit==1.38.0
google-api-python-client==2.149.0
groq==0.11.0
python-dotenv==1.0.1
pandas==2.2.3
httpx==0.27.0